dqn-pong/requirements.txt
numpy
pygame
tensorflow-gpu
